backend/requirements.txt
google-cloud-storage==3.3.0
google-cloud-video-transcoder
google-genai==1.31.0
tenacity
nest_asyncio
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv
python-multipart
ffmpeg-python
google-cloud-tasks